pyproject.toml
[tool.poetry]
name = "gps2024"
version = "0.1.0"
description = ""
authors = ["ArthurGolubev <gg.arthur.go@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.27.2"
requests-html = {extras = ["playwright"], version = "^0.10.0"}
lxml-html-clean = "^0.2.2"
pytest-playwright = "^0.5.2"
playwright = "^1.47.0"
python-slugify = "^8.0.4"
polars = "^1.7.1"
plotly = "^5.24.1"
pandas = "^2.2.3"
pyarrow = "^17.0.0"
pydantic = "^2.9.2"
kaleido = "0.2.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
